Rename the functions: 	add()		-> create_task() 	finished()	-> list_finished() 	didit()		-> mark_done() 	ls()		-> list_tasks() 	lscon()		-> list_contexts() 	lsprj()		-> list_projects() 	prio()		-> priority_change()
usage() exits 1 instead of 0.
Change usage() to reference arguments in long form even though the
script still accepts the 1-letter/2-letter arguments.
Make a remark in usage() that although case-insensitive matching works,
only lowercase queries can be fullfilled.

Remove main() and just put the case-esac at the end of the script.

Tasks are now created with c/create instead of a/add.
Tasks are appended via a/append instead of e a/edit append

Make this 0.8.2