Add set -e for safety in the beginning and set +e in the end.
Switch from -v to -q (verbose to quiet) mode since this seems to be the
saner default configuration.
Reflect this also in the man page.

Rename variables in a more consistent manner.

Note in the man page that in most cases only the first letter of the
command is needed.

Some minor fixes and clarifications.

Make this version 0.9.1.