Run $(date +%F) once in the beginning and then reuse the variable _date_today whenever needed.
Rework today(): Don't check for the +urgent project anymore. Instead
convert all due:<date> to unix time and check against _date_today to get
overdue tasks.
This changes the output in one significant way: Since the check is done
line by line and the output is immediate the overdue tasks are not
sorted by priority anymore but by date. This can be considered an
improvement since now both the priority and the time ordering is
immediately apparent by using both colour and ordering.