Add contrib/, move ticket(1) there
I recently stumbled upon Drew DeVault's blog post on contrib directories
(https://drewdevault.com/2020/06/06/Add-a-contrib-directory.html) which
introduced me to the concept.
As ticket(1) doesn't solve a real problem - you might as well just pipe
the output of todo(1) into grep(1) yourself - and including it into the
main script would mean a significant reshuffle of stuff, it seems a good
candidate to add to contrib/.
This also removes the manpage for ticket(1) since I only ever included
it for the sake of completeness. Now it is even less needed than before.