Get an accurate count of today's tasks
Counting n in list_today() before only provided info on whether anything
is due (n>0) or not (n=0), now the actual number of tasks due today is
counted. For now this doesn't provide additional benefit but it might
prove useful in the future. The cost is one grep(1) for every due date
in the past (overdue tasks) which should be fine.